requirements.txt
pandas~=2.0.3
openpyxl~=3.1.2
Unidecode~=1.3.6
python-dateutil~=2.8.2
xlrd~=2.0.1
